refractoring run_task
